services/analytics_service/pyproject.toml
[project]
name = "analytics_service"
version = "0.1.0"
description = "Thin HTTP enqueue API that pushes video-processing jobs onto the Redis queue"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115",
    "pydantic-settings>=2.5",
    "pyjwt>=2.9",
    "python-dotenv>=1.0",
    "redis>=5.0",
    "uvicorn>=0.30",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "httpx>=0.27",
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]
